+++ my-react-app/package.json
@@ -15,6 +15,7 @@
     "framer-motion": "^11.11.17",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-intersection-observer": "^9.13.1",
     "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
