+++ my-react-app/src/components/Footer.jsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+function Footer() {
+  return (
+    <footer className="bg-gradient-to-b from-transparent to-black/30 backdrop-blur-sm">
+      <div className="max-w-7xl mx-auto px-4 py-12">
+        <div className="grid grid-cols-1 md:grid-cols-4 gap-8">
+          <div className="space-y-4">
+            <h3 className="text-2xl font-bold bg-gradient-to-r from-white to-gray-400 bg-clip-text text-transparent">
+              The Metal Chef
+            </h3>
+            <p className="text-gray-400">
+              Pioneering Aluminum Excellence Since 1995
+            </p>
+          </div>
+          
+          <div className="space-y-4">
+            <h4 className="text-lg font-semibold text-white">Quick Links</h4>
+            <ul className="space-y-2">
+              <li><Link to="/" className="text-gray-400 hover:text-white transition">Home</Link></li>
+              <li><Link to="/about" className="text-gray-400 hover:text-white transition">About Us</Link></li>
+              <li><Link to="/products" className="text-gray-400 hover:text-white transition">Products</Link></li>
+              <li><Link to="/contact" className="text-gray-400 hover:text-white transition">Contact</Link></li>
+            </ul>
+          </div>
+          
+          <div className="space-y-4">
+            <h4 className="text-lg font-semibold text-white">Contact</h4>
+            <ul className="space-y-2 text-gray-400">
+              <li>123 Metal Street</li>
+              <li>Industrial District</li>
+              <li>contact@metalchef.com</li>
+              <li>+1 (555) 123-4567</li>
+            </ul>
+          </div>
+          
+          <div className="space-y-4">
+            <h4 className="text-lg font-semibold text-white">Follow Us</h4>
+            <div className="flex space-x-4">
+              {/* Add your social media icons/links here */}
+            </div>
+          </div>
+        </div>
+        
+        <div className="mt-12 pt-8 border-t border-white/10 text-center text-gray-400">
+          <p>© 2024 The Metal Chef. All rights reserved.</p>
+        </div>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
