+++ my-react-app/src/pages/About_us.css
@@ -1,6 +1,6 @@
 .about-container {
-  color: #ffffff;
-  padding: 4rem 2rem;
+  background: linear-gradient(to bottom, rgba(0,0,0,0.8), rgba(0,0,0,0.95));
+  padding: 6rem 2rem;
   min-height: 100vh;
 }
 
@@ -98,9 +98,8 @@
   flex: 1;
   min-width: 150px;
   padding: 1.5rem;
-  background: rgba(255, 255, 255, 0.05);
-  backdrop-filter: blur(8px);
-  border: 1px solid rgba(255, 255, 255, 0.1);
+  background: linear-gradient(145deg, rgba(255,255,255,0.05), rgba(255,255,255,0.02));
+  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
   border-radius: 1rem;
 }
 
@@ -132,8 +131,8 @@
 }
 
 .feature-item {
-  background: rgba(255, 255, 255, 0.05);
-  backdrop-filter: blur(8px);
+  background: linear-gradient(145deg, rgba(255,255,255,0.05), rgba(255,255,255,0.02));
+  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
   padding: 2rem;
   border-radius: 1rem;
   border: 1px solid rgba(255, 255, 255, 0.1);
